refactor : 이벤트 지원자 조회 로직 안에서 API 요청자가 이벤트 주최자가 맞는지 확인하는 로직 분리 (#85)
