+++ src/main/java/dnd/danverse/domain/matching/service/EventMatchComplexService.java
@@ -1,17 +1,12 @@
 package dnd.danverse.domain.matching.service;
 
-import static dnd.danverse.global.exception.ErrorCode.*;
 
 import dnd.danverse.domain.event.entitiy.Event;
 import dnd.danverse.domain.event.service.EventPureService;
 import dnd.danverse.domain.matching.dto.request.EventIdRequestDto;
-import dnd.danverse.domain.matching.dto.response.ApplicantsResponseDto;
 import dnd.danverse.domain.matching.entity.EventMatch;
-import dnd.danverse.domain.matching.exception.NotEventWriterException;
 import dnd.danverse.domain.profile.entity.Profile;
 import dnd.danverse.domain.profile.service.ProfilePureService;
-import java.util.List;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -71,28 +66,6 @@ public class EventMatchComplexService {
   }
 
 
-  /**
-   * 1. 입력받은 이벤트 아이디로 이벤트글을 가져온다. (event)
-   * 2. 이벤트 글에 있는 프로필을 가져온다. (writer)
-   * 3. 입력받은 멤버 아이디를 통해 프로필을 가져온다. (profile)
-   * 4. profile 과 writer 를 비교하여 다르면 403에러 발생한다.
-   * 5. 동일하면 신청자 리스트를 가져와서 dto 에 담아 반환한다.
-   *
-   * @param eventId 신청자 리스트를 조회하고자 하는 이벤트 글 아이디.
-   * @param memberId 신청자 리스트를 조회하려고 하는 사용자 Id.
-   */
-  public List<ApplicantsResponseDto> getApplicants(Long eventId, Long memberId) {
-    Event event = eventPureService.getEvent(eventId); //입력받은 이벤트아이디로 이벤트글을 찾아온다.
-    Profile writer = event.getProfile(); //가져온 이벤트안에 있는 프로필의 아이디를 가져온다.
-    Profile profile = profilePureService.retrieveProfile(memberId); //프로필 순수 서비스에서 입력받은 멤버의 아이디의 프로필을 가져온다.
-
-    if (writer.isNotSame(profile)) { //프로필끼리 비교하여 틀리면 403에러 발생
-      throw new NotEventWriterException(EVENT_MATCH_NOT_WRITER);
-    }
-    List<EventMatch> eventMatches = eventMatchPureService.getApplicants(eventId);
-
-    return eventMatches.stream().map(ApplicantsResponseDto::new).collect(Collectors.toList());
-  }
 
 
 }
